requirements.txt
aws-cdk-lib==2.49.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-batch-alpha
aws-cdk.aws_lambda_python_alpha